+++ src/index.ts
@@ -1,22 +1,13 @@
-// layouts
-export { default as DefaultLayout } from '@/package/layouts/defaultLayout.vue';
-export { default as LeftSidebarLayout } from '@/package/layouts/leftSidebarLayout.vue';
 // components
-export { default as Demo } from '@/package/components/demo.vue';
 export { default as Link } from '@/package/components/link.vue';
 export { default as Card } from '@/package/components/card.vue';
-export { default as Logo } from '@/package/components/logo.vue';
 export { default as Menu } from '@/package/components/menu.vue';
-export { default as Header } from '@/package/components/header.vue';
-export { default as Footer } from '@/package/components/footer.vue';
 export { default as Button } from '@/package/components/button.vue';
 export { default as Input } from '@/package/components/input.vue';
 export { default as Select } from '@/package/components/select.vue';
 export { default as Checkbox } from '@/package/components/checkbox.vue';
 export { default as Switch } from '@/package/components/switch.vue';
-export { default as Content } from '@/package/components/content.vue';
 export { default as Popover } from '@/package/components/popover.vue';
-export { default as Container } from '@/package/components/container.vue';
 // config
 export { default as VueForge, setTheme, updateTheme, getTheme } from '@/package/config/index';
 // themes