+++ src/package/themes/default/components/checkbox.ts
@@ -2,6 +2,7 @@ export default {
     size: '1rem',
     gap: '0.5rem',
     borderRadius: '4px',
+    checkBorderRadius: '2px',
     borderColor: 'var(--vf-border-color)',
     backgroundColor: 'var(--vf-bg-color)',
     checkedBackgroundColor: 'var(--vf-blue)',