+++ src/package/themes/default/components/footer.ts
@@ -1,8 +1,0 @@
-export default {
-    color: 'var(--vf-secondary-text-color)',
-    padding: '12px',
-    marginTop: '24px',
-    borderTop: '1px solid var(--vf-footer-border-color)',
-    borderColor: 'var(--vf-border-color)',
-    backgroundColor: 'var(--vf-bg-color)',
-};