+++ src/package/themes/default/components/header.ts
@@ -1,8 +1,0 @@
-export default {
-    height: '55px',
-    padding: '0',
-    borderColor: 'var(--vf-border-color)',
-    borderBottom: '1px solid var(--vf-header-border-color)',
-    marginBottom: '24px',
-    backgroundColor: 'var(--vf-bg-color)',
-};