+++ src/package/themes/default/components/menu.ts
@@ -1,4 +1,8 @@
 export default {
+    iconGap: '6px',
+    submenuOffset: '12px',
+    separatorThickness: '1px',
+    separatorHeight: '1rem',
     separatorColor: 'var(--vf-border-color)',
     link: {
         hoverColor: 'var(--vf-blue)',