+++ src/package/themes/default/components/switch.ts
@@ -3,6 +3,8 @@ export default {
     height: '1.25rem',
     thumbSize: '1rem',
     gap: '0.5rem',
+    thumbOffset: '2px',
+    thumbTranslateOffset: '4px',
     backgroundColor: 'var(--vf-border-color)',
     checkedBackgroundColor: 'var(--vf-blue)',
     thumbColor: '#ffffff',